Android Journey Update Back Stack 🐞Fixed
Not Fixed Actually But, Activity is Exited After the fragment loads giving an illusion that when back button is pressed it redirects to list instead of view page. The bug was created because of Parent Activity on Manifest File from [android:parentActivityName=".views.activities.home.HomeActivity"] line of code which provides up action button.